One last fix for incorrectly type mask literals (32/64 bit issue).

+++ ext/bitset/bitset.c
@@ -156,7 +156,7 @@ static VALUE rb_bitset_cardinality(VALUE self) {
     for(i = 0; i < max; i++) {
         uint64_t segment = bs->data[i];
         if(i+1 == max)
-            segment &= ((1 << (bs->len & 0x3F)) - 1);
+            segment &= ((((uint64_t) 1) << (bs->len & 0x3F)) - 1);
         count += __builtin_popcountll(segment);
     }
     return INT2NUM(count);
